requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.29.7
toml==0.10.2
requests==2.31.0
apache-airflow==2.7.3
pyspark==3.5.0
pandas==2.1.3
pyarrow==14.0.1
email-validator>=2.0.0
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
vnstock>=1.0.0
mplfinance>=0.12.9b0
matplotlib>=3.7.0
plotly>=5.17.0
# TradingAgents dependencies
typing-extensions
langchain-openai
langchain-experimental
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
tqdm
pytz
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai
python-dotenv
paho-mqtt

